+++ cmd/root.go
@@ -20,16 +20,7 @@ var (
 // RootCmd represents the base command when called without any subcommands
 var RootCmd = &cobra.Command{
 	Use:   "lakuh",
-	Short: "A brief description of your application",
-	Long: `A longer description that spans multiple lines and likely contains
-examples and usage of using your application. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
-	// Uncomment the following line if your bare application
-	// has an action associated with it:
-	//	Run: func(cmd *cobra.Command, args []string) { },
+	Short: "Distributed radio automation.",
 }
 
 // Execute adds all child commands to the root command and sets flags appropriately.
@@ -51,7 +42,7 @@ func Execute() {
 func init() {
 	cobra.OnInitialize(initConfig)
 
-	RootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.lakuh.yaml)")
+	RootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.lakuh.toml)")
 	RootCmd.PersistentFlags().BoolVar(&debug, "debug", false, "show debug messages")
 
 	RootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
@@ -66,8 +57,9 @@ func initConfig() {
 		// Find home directory.
 		home, err := homedir.Dir()
 		if err != nil {
-			fmt.Println(err)
-			os.Exit(1)
+			// fmt.Println(err)
+			// os.Exit(1)
+			logrus.Fatal(err)
 		}
 
 		// Search config in home directory with name ".lakuh" (without extension).
@@ -77,20 +69,20 @@ func initConfig() {
 
 	viper.AutomaticEnv() // read in environment variables that match
 
-	// If a config file is found, read it in.
-	if err := viper.ReadInConfig(); err == nil {
-		fmt.Println("Using config file:", viper.ConfigFileUsed())
+	if err := viper.ReadInConfig(); err != nil {
+
+		logrus.Debug(err)
+
+		if e := generateDefaults(); e != nil {
+			logrus.Fatal(e)
+		}
+
 	}
 
 	defaults()
 }
 
 func defaults() {
-
-	viper.SetDefault("lakuh.database", "~/.lakuh/db/main.db")
-	viper.SetDefault("audio.storage", "~/.lakuh/audio")
-
-	///////////////////
 
 	storage, err := utils.FullPath(viper.GetString("audio.storage"))
 	if err != nil {